Création du dossier modeles et déplacement du fichier donnees.py dedans
